Updated the IUP to 3.10.1, the libcsoup to 0.9.2. Using a local makefile to manage the upsource libraries. Cleaned up the libgd.

+++ external/iup/srcmatrixex/iupmatex_copy.c
@@ -141,7 +141,7 @@ static int iMatrixExStrGetInterval(const char* interval, int num_lin, char* sele
   memset(selection, '0', num_lin+1);
   *selection_count = 0;
 
-  len = strlen(interval);
+  len = (int)strlen(interval);
 
   do
   {
